Starter Code for Game Board
Starter code for the game board contained in "board.py". When executed it will create an empty game board and then wait for the file "TEST.go" to be present in the current directory. Once found it will load the file "move_file.txt", parse though it and add a piece for the corresponding team to the board. Then the entire board is printed out to the console. All the functions are commented with brief descriptions of their function. The “Main()” function is called when the code is executed.