Min Max with Alpha Beta Pruning
Min Max with Alpha Beta Pruning implemented in the file "algorithm.py". In the file is the class "Node" which is a single node used to construct a tree of nodes. To preform a Min Max search, input the desired node to search into min_max_alpha_beta() and the function will return the best of the child nodes of the inputted node.