Updated Min Max with depth limits, and timouts
min_max_alpha_beta() now must also take in a depth limit else there will be no limit and the search wil go to the bottom of the tree. Additionally a timeout function has been implemented where min_max_alpha_beta() will return the best path it was able to find after a certain amount of seconds has passed. The amount of seconds until timeout is stored in the constant TIMEOUTAMOUNT as an int.